KFModTool: Move ent. class & instance to table models, misc optimizations
This huge commit switches the entity instance and class tables in the
map editor from QTableWidget to QTableView widgets. To do this, table
models for the entity class and instance have been created.

Using the model/view pattern enables us to write cleaner, more efficient
and more maintainable code and also allows us to make the tables
prettier.

This commit also brings updates to the game structures as well as other
optimizations.