KFModTool and tfile_tool: Improve PrettyNamer design
The PrettyNamer class now holds richer names for the files, that are
stripped down (that is, have any special characters removed) when
exporting the files but can be used in their fully readable glory in
KFModTool. This is useful not only for integration purposes, but it also
allows us to add newly discovered names to the PrettyNamer class and
have the change apply to both KFModTool and tfile_tool.