KFModTool: Prevent error message spam by keeping the two last errors
KFMTError now has a queue that keeps track of the last two error
messages shown by the error function. When the error function is
triggered, KFMTError now checks if the given error message is in the
queue. If it is, the message is instead logged to stderr without a
messagebox. Otherwise, the messagebox is shown normally.