KFModTool: Move PrettyNamer and TFile to common dir to avoid symlinks
Symbolic links aren't handled too well on Windows, so this commit moves
the PrettyNamer and TFile classes, used by both KFModTool and tfile_tool
to a newly-created "common" directory. This allows us to share the files
without the need for symlinks and should provide a more pleasant
experience to Windows users who want to try hacking the tools.