KFModTool: Add support for more games (1J, 1J demo, 1EU, 2J/3U/PS)
This is a huge commit that adds support for loading files from other
King's Field games. It adds support for the MIX file format used in 1J
and prepares the tool to load files from King's Field 1J, the demo for
King's Field 1J from the "DemoDemo PlayStation Soukan-gou Vol. 1" CD,
the European King's Field (2J/1U), as well as both King's Field 3J (2U)
and King's Field III: Pilot Style.

This commit does not yet re-implement the ability to save changes, so
this snapshot is more of a KFViewTool than a KFModTool.