KFModTool: Beginnings of Shadow Tower model support
This commit adds a lot of work done by @TheStolenBattenberg and me to
support loading Shadow Tower's models properly. We have found that
Shadow Tower has two different MO revisions and a customized TMD
contained within them that stores primitives in a different way to save
space. For now, animation is not implemented for these models as it will
require specific coding for the different MO revisions.

The Shadow Tower TMD format differs from standard TMD in that packets do
not have individual headers containing mode, flag, ilen and olen.
Instead, after the primitive count in the TMD header, they store 8
unsigned 16-bit integers that represent the amount of primitives of a
certain mode. This allows the model files to save 4 bytes per primitive.

As of this commit, @TheStolenBattenberg has found evidence that points
to the primitive counters representing different primitive types as
follows:

    ushort numFT3_24;        //Size: 0x14
    ushort numGT3_34;        //Size: 0x18
    ushort numFT4_2c;        //Size: 0x1c
    ushort numGT4_3c;        //Size: 0x20
    ushort numF3_20;        //Size: 0xc
    ushort numG3_30;        //Size: 0x10
    ushort numF4_28;        //Size: 0x10
    ushort numG4_38;        //Size: 0x14

Some of the models look garbled and/or crash KFModTool at the moment,
and we are investigating possible reasons for this happening.