KFModTool: A huge refactor!
This update has been in the works for quite a while. It finally brings
some actual architecture to KFMT, transforming it from a spaghetti mess
of code thrown in the main window to a properly structured project. Now a
class called KFMTCore is responsible for... well... *core* functionality
such as loading files, saving and such. It has a single global instance,
which means that any data handler can load an arbitrary game file whenever
it wants.

Oh yeah, data handlers. Stuff like the Map and TextureDB classes are now
subclasses of the KFMTDataHandler class. It's pretty much just an interface
to standardize basic behavior across the different data handlers, and
should allow us to do some pretty nifty stuff in the future. Right now its
main use is enabling us to standardize the KFMTEditor superclass.

KFMTEditor is a superclass that standardizes editor widgets such as the map
editor, the texture editor, etc. This makes it easier for us to add new
editors without having to write too much boilerplate code, and gives us
some constraints which help remind us to do the right things.

Directory structure is likely to change a bit still as I'll have to see just
how many of our datahandlers and models are completely KF2-specific. I know
that we can re-use a lot of that stuff for KF3 and such, but I'll still have
to investigate.

As for actual functionality, this KFMT update also has that covered!
* 3D models now have textures! And when they don't, they look terrible. I'll
  see about fixing this in the next release.
* The map editor has been revamped, with proper understanding of some of
  the tile flags, a tile preview and a better system for editing entities,
  objects and VFxs. Unfortunately, due to this current design, entity
  editing gets tight on smaller screens...
* 3D map previewing! Now you can browse through maps in 3D! Use WASD to move
  and to rotate the view, hold the middle mouse button (starting on the 3D
  viewer) and rotate your mouse!
* Game EXE editing for KF2U. This allows you to edit the strings and mess with
  shop stuff for now. More functionality will come later :)

There are still known issues with this release, but I had to get it out.
Otherwise, I'd be sitting on it forever.

Hope y'all enjoy!