KFModTool: Huge update!
This monster commit brings a slew of internal changes to KFMT to try and make it easier to support other games.

The huge kf2.h/cpp file pair has been refactored into smaller, domain-specific files that are now located in the kf2 folder.
This should improve compile times, given that now parts of the program that deal with a specific type of data (e.g. entity, object stuff) now only needs to include that specific header file, which avoids repeating parsing of a huge header file.

Data models have also been refactored to be more concised, and might in the future be split along game folders as well to facilitate understanding.

Apart from those changes, other pieces of overall jank have been improved upon, though there still is a lot to work on - let's hope that I can avoid huge commits like this one in the future.

And also hope that GitHub Actions is able to compile this without issues.